move some logic from hook back into component
